Moved b64 functions to their own provider; Refactored reactions after change; added inputs to register function to handle profile pic image upload; wrapped everything in ImageProvider
