+++ src/components/auth/Register.js
@@ -1,5 +1,6 @@
-import React, { useRef } from "react"
+import React, { useContext, useRef, useState } from "react"
 import { Link, useHistory } from "react-router-dom"
+import { ImageContext } from "../images/ImageProvider"
 import "./Auth.css"
 
 export const Register = (props) => {
@@ -12,6 +13,15 @@ export const Register = (props) => {
     const passwordDialog = useRef()
     const username = useRef()
     const history = useHistory()
+    const {createImageString, b64, setB64} = useContext(ImageContext)
+
+    // const [profileImage, setProfileImage] = useState("")
+
+    // const handleInputChange = (event) => {
+    //     let imageCopy = {...profileImage}
+    //     imageCopy[event.target.id] = event.target.value
+    //     setProfileImage(imageCopy)
+    // }
 
     const handleRegister = (e) => {
         e.preventDefault()
@@ -24,7 +34,7 @@ export const Register = (props) => {
                 "email": email.current.value,
                 "password": password.current.value,
                 "bio": bio.current.value,
-                "profile_image_url": "",
+                "profile_image_url": b64,
             }
 
             return fetch("http://127.0.0.1:8000/register", {
@@ -42,6 +52,7 @@ export const Register = (props) => {
                         history.push("/")
                     }
                 })
+                .then(setB64(""))
         } else {
             passwordDialog.current.showModal()
         }
@@ -83,6 +94,11 @@ export const Register = (props) => {
                 <fieldset>
                     <label htmlFor="verifyPassword"> Verify Password </label>
                     <input ref={verifyPassword} type="password" name="verifyPassword" className="form-control" placeholder="Verify password" required />
+                </fieldset>
+                <fieldset>
+                    <label htmlFor="reactionLabel">Upload a Profile Image </label> <br></br>
+                    {/* <input type="text" id="image" onChange={handleInputChange} value={profileImage}></input><br></br> */}
+                    <input type="file" id="profile_image_url" onChange={createImageString} />
                 </fieldset>
                 <fieldset style={{
                     textAlign: "center"
