+++ src/components/reaction/ReactionProvider.js
@@ -4,7 +4,6 @@ export const ReactionContext = createContext()
 
 export const ReactionProvider = props => {
 
-    const [b64, setB64] = useState("")
     const [reactions, setReactions] = useState([])
 
     const getReactions = () => {
@@ -15,18 +14,6 @@ export const ReactionProvider = props => {
       })
       .then(res => res.json())
       .then(setReactions)
-    }
-
-    const getBase64 = (file, callback) => {
-        const reader = new FileReader();
-        reader.addEventListener('load', () => callback(reader.result));
-        reader.readAsDataURL(file);
-    }
-    
-    const createReactionImageString = (event) => {
-        getBase64(event.target.files[0], (base64ImageString) => {
-            setB64(base64ImageString)
-        });
     }
 
     const createReaction = reaction => {
@@ -43,7 +30,7 @@ export const ReactionProvider = props => {
     }
 
     return (
-        <ReactionContext.Provider value={{ createReactionImageString, b64, createReaction, getReactions, setB64, reactions }} >
+        <ReactionContext.Provider value={{createReaction, getReactions, reactions }} >
             { props.children }
         </ReactionContext.Provider>
     )
